Traceability graph improved, user stories names and minor fixes
Traceability graph improved a lot, nodes are now organized in columns, links are curved and there's an effort to make links less cluttered.
Overview of changes added
User story name field is now formed at save to be able to be shown when it's called as a generic artifact.
General progress now ponder tests completion as 50% of its own completion instead of the previous 100%.